Add Apollo server - Add the posibility of switch between Apollo and Express servers
