+++ package.json
@@ -10,6 +10,7 @@
   "author": "Hamid Javadi",
   "license": "ISC",
   "dependencies": {
+    "apollo-server": "^3.9.0",
     "express": "^4.18.1",
     "express-graphql": "^0.12.0",
     "graphql": "^16.5.0",
