+++ src/config.ts
@@ -1,0 +1,3 @@
+export const config = {
+  useExpress: true,
+};
