+++ src/server-apollo.ts
@@ -1,0 +1,8 @@
+import { ApolloServer } from "apollo-server";
+import schema from "./schema";
+
+const apolloServer = new ApolloServer({
+  schema,
+});
+
+export default apolloServer;
