+++ src/type.defs.ts
@@ -1,0 +1,14 @@
+const typeDefs = `
+    type User {
+        id: ID
+        name: String
+        family: String
+        email: String
+    }    
+
+    type Query {
+        profile: User
+    }
+`;
+
+export default typeDefs;
